chore: migrate mapObject and other stuffs to utils
