+++ packages/tailwind-config/src/util.ts
@@ -66,100 +66,6 @@ Got: ${JSON.stringify(Array.from(effectKeys))}`)
   }
 }
 
-/**
- * @example
- * ```js
- * mapKeys({ a: 'aa', b: 'bb' }, (key) => key.toUpperCase()) // => { A: "aa", B: "bb" }
- * ````
- */
-export function mapKeys<V, K extends string>(
-  object: Record<string, V>,
-  callback: (key: string) => K
-) {
-  return Object.fromEntries(
-    Object.entries(object).map(([key, value]) => [callback(key), value])
-  ) as Record<K, V>
-}
-
-/**
- * @example
- * ```js
- * mapObject({ a: 'aa', b: 'bb', c: 'cc' }, (key, value) =>
- *   key === 'b' ? undefined : [key + '1', value.toUpperCase()]
- * ) // => { a1: "AA", c1: "CC" }
- * ```
- */
-export function mapObject<
-  SourceKey extends string,
-  SourceValue,
-  DestKey extends string,
-  DestValue
->(
-  source: Record<SourceKey, SourceValue>,
-  callback: (
-    key: SourceKey,
-    value: SourceValue
-  ) => [DestKey, DestValue] | null | undefined
-) {
-  return Object.fromEntries(
-    Object.entries(source).flatMap(([key, value]) => {
-      const entry = callback(key as SourceKey, value as SourceValue)
-      if (entry) {
-        return [entry]
-      } else {
-        return []
-      }
-    })
-  ) as Record<DestKey, DestValue>
-}
-
-/**
- * @example
- * ```js
- * flatMapObject({ a: 'aa', b: 'bb' }, (key, value) => [
- *   [key + '1', value + '1'],
- *   [key + '2', value + '2'],
- * ]) // => { a1: "aa1", a2: "aa2", b1: "bb1", b2: "bb2" }
- * ```
- */
-export function flatMapObject<
-  SourceKey extends string,
-  SourceValue,
-  DestKey extends string,
-  DestValue
->(
-  source: Record<SourceKey, SourceValue>,
-  callback: (key: SourceKey, value: SourceValue) => [DestKey, DestValue][]
-) {
-  return Object.fromEntries(
-    Object.entries(source).flatMap(([key, value]) => {
-      return callback(key as SourceKey, value as SourceValue)
-    })
-  ) as Record<DestKey, DestValue>
-}
-
-/**
- * @example
- * ```ts
- * filterObject(
- *   { a: 'aa', b: 'bb', c: 'cc' },
- *   (value): value is string => value !== 'bb'
- * ) // => { a: "aa", c: "cc" }
- * ```
- */
-export function filterObject<Source, Dest extends Source>(
-  source: Record<string, Source>,
-  fn: (value: Source) => value is Dest
-) {
-  return mapObject(source, (key, value) => {
-    if (fn(value) === true) {
-      return [key, value]
-    } else {
-      return null
-    }
-  }) as Record<string, Dest>
-}
-
 export function camelToKebab(value: string) {
   return value
     .replace(/(?<small>[\da-z]|(?=[A-Z]))(?<capital>[A-Z])/gu, '$1-$2')
